Move Nunjucks renderer into common
