+++ app/common/nunjucks/filters/index.js
@@ -1,0 +1,8 @@
+/**
+ * Nunjucks filters
+ */
+const { componentNameToMacroName } = require('../../../../lib/helper-functions.js')
+
+module.exports = {
+  componentNameToMacroName
+}
