+++ app/common/nunjucks/globals/index.js
@@ -1,0 +1,8 @@
+/**
+ * Nunjucks globals
+ */
+const markdown = require('marked')
+
+module.exports = {
+  markdown
+}
